Move code to utils.py
