+++ setup.cfg
@@ -1,2 +1,2 @@
 [pycodestyle]
-ignore = E501
+ignore = E501, E702
